+++ CHANGELOG.md
@@ -5,4 +5,5 @@ Changed:
 - CRN.equilibrate now accepts 2D initial conditions with specialized CRN.params
 
 Fixed:
+- ParameterMap.specification_for behaves better with absent species
 - CRN.parametrize_for is now more careful when substituting specialized parameters in expressions